date로 db에 변동을 줄시 다음 작업을 하지 않게하는 변수 flag_date추가 및 코드 적용

+++ django_server/recommendation/views.py
@@ -228,6 +228,7 @@ def reservation(request):
     connection.commit()
     temp = cursor.fetchall()
     min_date=0
+    flag_date=0
     for i in temp:
         if i[4] <min_date:
             min_date=i[4]
@@ -239,22 +240,22 @@ def reservation(request):
         strSql = 'INSERT INTO preference_table(reservation_user, seat_code , count ,date,score,density) VALUES ((%s),(%s),(%s),(%s),(%s),(%s))'
         cursor.execute(strSql, (userid, seat_code, 1, date, score, density,))
         connection.commit()
-
-    for i in temp:
-        if i[2] == seat_code:
-            temp_count=i[3]+1
-            strSql='UPDATE preference_table SET  count =(%s), score=(%s),density=(%s) WHERE reservation_user=(%s) and seat_code=(%s)'
-            cursor.execute(strSql, (temp_count,score,density,userid,seat_code,))
-            connection.commit()
-    for i in temp:
-        if i[5] < score:
-            strSql='DELETE FROM preference_table WHERE reservation_user=(%s) and seat_code=(%s)'
-            cursor.execute(strSql, (userid, i[2] ,))
-            connection.commit()
-            strSql = 'INSERT INTO preference_table(reservation_user, seat_code , count ,date,score,density) VALUES ((%s),(%s),(%s),(%s),(%s),(%s))'
-            cursor.execute(strSql, (userid, seat_code , 1 ,date ,score ,density, ))
-            connection.commit()
-
+        flag_date=1
+    if flag_date !=1:
+        for i in temp:
+            if i[2] == seat_code:
+                temp_count=i[3]+1
+                strSql='UPDATE preference_table SET  count =(%s), score=(%s),density=(%s) WHERE reservation_user=(%s) and seat_code=(%s)'
+                cursor.execute(strSql, (temp_count,score,density,userid,seat_code,))
+                connection.commit()
+        for i in temp:
+            if i[5] < score:
+                strSql='DELETE FROM preference_table WHERE reservation_user=(%s) and seat_code=(%s)'
+                cursor.execute(strSql, (userid, i[2] ,))
+                connection.commit()
+                strSql = 'INSERT INTO preference_table(reservation_user, seat_code , count ,date,score,density) VALUES ((%s),(%s),(%s),(%s),(%s),(%s))'
+                cursor.execute(strSql, (userid, seat_code , 1 ,date ,score ,density, ))
+                connection.commit()
     strSql = "SELECT * FROM preference_table where reservation_user = (%s)"
     cursor.execute(strSql, (userid,))
     connection.commit()
